+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 ### Added
 
 - `with_innerhtml` parameter for matching blocks by their inner HTML.
+- `ancestor_of` parameter for matching blocks by their inner blocks.
 - `has_innerblocks` parameter for matching blocks by whether they have inner blocks. Includes companion `\Alley\WP\Validator\Block_InnerBlocks_Count` validator,
 - Passing a single block instance will return matches within its inner blocks.